==== Just Unzip ==== Tags: unzip, upload, multiple, admin, post Contributors: jrevillini, Vincent Blavet
Just Unzip (LGPL Licensed) takes a zip file you upload from the "Write" page, unzips it, stores the zipped files in your upload folder, and associates the unzipped files with the current post.    Just Unzip makes use of the PclZip php library (LGPL Licensed).

inline-uploading.php - a replacement for wp/wp-admin/inline-uploading.php which is the document which loads into the uploading iframe when you post.  The difference between mine and the built-in file is that mine will split a zip archive into individual files and attach them to the current post.

just-unzip.php - the main plugin file which contains the just-unzip class

pclzip.lib.php - the library used for zip/unzip functions

readme.txt - to comply with wp plugin dev guidelines, a text synopsis of the plugin, its tags, and contributors.


git-svn-id: https://plugins.svn.wordpress.org/just-unzip/trunk@6543 b8457f37-d9ea-0310-8a92-e5e31aec5664